refactor: type artifact ownership
